ììDB데이ã터로 지도 정보 불러오기 성공

+++ src/main/java/com/doNotWorry/FoodDatas.java
@@ -25,9 +25,9 @@ public class FoodDatas {
 
     String storePhNum;
 
-    Float latitude;
+    double latitude;
 
-    Float longitude;
+    double longitude;
 
     List<String> menu;
